java

+++ SpringCore/src/main/java/spring/core/session04/proxy/sta/Calc.java
@@ -1,5 +1,5 @@
 package spring.core.session04.proxy.sta;
 
 public interface Calc {
-	double div(int x, int y); // 除法計算
+	int div(int x, int y); // 除法計算
 }
